+++ CHANGELOG.md
@@ -20,6 +20,7 @@ limitations under the License.
 
 - new: TensorRT Timing Tactics Cache Management - using timing tactics cache files for optimization performance improvements
 - new: Added throughput saturation verification in `nav.profile()` (enabled by default)
+- new: Allow to override Inplace cache dir through `MODEL_NAVIGATOR_DEFAULT_CACHE_DIR` env variable
 - fix: torch dynamo export and torch dynamo onnx export
 - fix: measurement stabilization in `nav.profile()`
 - fix: inplace inference through Torch